refactor(batch-update): divide service
batch-update service was divided on checkers and assignments makers
